item-eqv?
Now that items created with make-item contain a 'digest item-ref and
items retrieved by the Backing Store contain an 'opaque item-ref we need
a way to check whether two items are "equivalent" without their
internal representations being strictly "equal".

We reuse the -equal? and -eqv? nomenclature from Scheme without it being
strictly the same. item-eqv? returns #t if the items have the same blob
contents *and* their item-refs are either also exactly the same *or*
their item-refs are of different types. i.e. item-eqv? usually* (see
note) returns #t if the items have the same contents. It always returns #f if the items have different contents. Careful readers will notice that this is more like the Java comparison semantics than the Scheme ones. i.e. things that are eqv? are not always equal? but things that are equal? are always eqv?.

In contrast, the existing item-equal? returns #t iff their contents
*and* their item-ref are exactly the same. This means that equivalent
items with different refs will get #f from item-equal?

If we supported more than one algorithm for 'digest item-ref or more
than one Backing Store it would be possible for a pair of equivalent
items to have item-refs that were of the same type and have different
type specific information. In this case, item-eqv? would give the
incorrect answer (returning #f). The situations in which we should
return #t or #f depend on how those not-yet-existent things are
implemented so we err on the side of caution by never delcaring items
equivalent if they might not be.

* Note: we say "usually" because of the caveat about support for
'digest item-refs of differentand 'opaque. In the implementation as it
stands today, items with the same contents will always get #t from
item-eqv?

I really should fix this nice mess. I'll do that in Future Work.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>