Introduce 'opaque item-refs
'opaque item-refs contain a reference to a specific item in the Backing
Store. This allows the Backing Store to use its own naming and
de-duplication schemes for the items it stores and do management of the
digest information separately.

Here we rework the item-ref internal format so that it is more regular
between each of the different types of item-ref. The format is as
follows:

`(item-ref <type> <type-specific-data>)

In the case of a 'digest type item-ref the type-specific-data is as
follows:

`(<algorithm> <digest>)

In the case of an 'opaque type item-ref the type-specific-data should be
Backing Store dependent. For now we're intending to use SQLite row-ids
to name the items. Therefore, the type-specific-data is as follows:

`(<item-id>)

The make-item-ref procedure is exported so we do not teach it how to
construct 'opaque item-refs. For that we supply make-item-ref-opaque and
designate it for the sole use of the Backing Store code. Looking up
items in the Backing Store should be done through the use of either 'digest
item-refs or 'opaque item-refs that have previously been returned by the
Backing Store. An 'opaque item-ref is only guranteed to remain valid for
the lifetime of the item-ref object and therefore, as the name suggest,
we do not provide any way for the user to interact with them:
item-ref-item-id is for the use of Backing Store code only.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>