Teach the SQLite Backing Store how to persist Register objects
This patch introduces register-store-add!

We don't currently use any transactions because we require SQLite's
SAVEPOINT transaction model and that is not currently available through
the sql-de-lite egg. As this code only uses a single statement to
achieve its goal this does not lead to a correctness issue for the
operation of allocating a Register. However, if something later goes
wrong, the Register creation cannot be rolled back and the contents of
the Register may be left in an inconsistent state.

Moreover, the inability to rollback the Register creation means that the
namespace for naming Registers can become polluted.

We don't yet provide a way to retrieve a Register from the Backing Store
by name. For now we rely on the user to keep hold of their own register
objects and use them appropriately. This will be fixed in future work.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>