Rework the register object to use the SQLite Backing Store
The register object needs to keep track of a lot less state than before!
Now we just need a reference that the Backing Store can use to find it
and a version number stating how many entrys in the entry log we want to
consider when operating on the Register.

The version number ensures that others who have references to a Register
do not see it change underneath them. It also allows the Register to be
opened at any point in its history, although non-latest version may not
be writable.

Due to the requirements of the Backing Store interface, the isolation
properties given by the version number apply even if the Backing Store
itself has no support for transactional operations.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>