Teach the Backing Store how to persist entries
Here we provide entry-store-add.

entry-store-add returns a new Register object that reflects the addition
of the supplied entry onto the end of the entry log for the supplied
Register.

Unlike item-store-add!, item-store-add-digest! and register-store-add!,
entry-store-add does not end with '!' because its effects are not
visible to anyone other than the recipient of the return value.

Because this Backing Store only allows the latest version of a Register
to be written, a successful call to entry-store-add will render all
other references to the Register read-only.

We don't currently use any transactions because we require SQLite's
SAVEPOINT transaction model and that is not currently available through
the sql-de-lite egg. As entry-store-add uses a number of statements to
complete its work this means that the behaviour of concurrent calls to
entry-store-add are be undefined (even if the db-ctx threading situation
were resolved).

It is not possible to make the entry-store-add operation safe without
proper database transactions. We have two options available to us,
neither of which lead to safety:

Firstly, we could call insert-entry before insert-entry-items. This
would allow us to reserve the next entry number and therefore exclude
other writers but, concurrent readers will be able to see our incomplete
entry. Should the later operations fail, it could lead to the Register
being in in an inconsistent, undetectable and therefore, unrecoverable
state.

Secondly, we could call insert-entry-items before insert-entry. This
would mean that we would race with other writers to claim the entry
number and any concurrently written entrys would end up with an
undefined item-list state. However, the entry would not be visible to
other readers until the entry-number race had been won!

We pass the backing-store-ref: argument to update-register (even though
it doesn't change) to reflect the idea that a Backing Store that allows
non-latest versions of a Register to remain writeable might modify this
piece of data.

In the database we store dates in UNIX-seconds-since-the-epoch format.
This means that they must be in UTC. It is not possible to reliably
convert dates in the future to UTC because the timezone rules may change
between now and that date. Rather than enforce a policy that all
timestamps must be in the past, date->integer enforces that all dates
must be in UTC. This way the user is responsible for performing the
timezone conversion themselves. RSF already requires all dates to be in
UTC so we wrap srfi-19's string->date procedure to ensure that the
srfi-19 dates we end up with after processing an RSF file with read-rsf
remain in UTC no matter what timezone the machine is in.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>