entry-store-{key-ref,keys} to read entrys from the Backing Store by key
This will be used to query the Records that the Register contains.  A
Record is the latest entry in the Register that contains the specified
key in its key field (but see details about "tombstones" below).

An entry contains some fields of its own and also a list of zero or more
items. This is represented in the Backing Store as an `entry` table
containing the entry fields and an `entry-items` join table that
connects the entry to its items. We need to do some SQL gymnastics to
work out the lastest `entry-number` in the `entry` table for a given key
and then join that back to the `entry` table to get the rest of the
entry fields.  We then join onwards to the `entry-item` table to get the
`item-id`s for that `entry-number` and then join on again to the `items`
table to get the contents of the relevant items.

As the entry-store-{key-ref,keys} procedures return fully resolved
items, we neccessarily have to deal with both entry and item concerns in
this part of the code.

The hierarchical nature of the entrys and their items are not elegantly
expressed in a flat, table based SQL result set. We use the traditional
method of a LEFT OUTER JOIN so that we receive at least one row for each
and every entry. If that entry has no items then the item columns from
the JOIN turn out NULL. Otherwise, we receive one row per item per
entry.

`wide-entry-item-rows->entries` executes the relevant query and keeps
track of the inter-row state, calling make-item and make-entry as
appropriate: when the key in the entry columns of the result set changes
then it's time for a new entry, otherwise we've got another item for the
existing entry. Whilst neither of entry-store-{key-ref,keys} request a
specific ordering, this code preserves the order in which the database
returns things.

entry-store-key-ref takes one or two keys and returns all the entrys
with keys that fall lexographically (inclusive) between the two. This
allows higher layers to perform range queries over the Record Set. If
only one key is provided this is the same as a point query for that
particular key and a single or zero row result is expected.

entry-store-keys returns an entry for every key in the Record Set. This
could be a lot of data and it'd probably be wise to refactor this so
that it could be accessed with an iterator or lazy-seq style interface.
It meets our needs for now tho'.

Both entry-store-{key-ref,keys} (and anything else using
`wide-entry-item-rows->entries) have full support for entry tombstones.
That is, an entry that contains an empty list of items is treated as a
"tombstone" and that key is removed from the Record Set. This allows
things to be deleted from the Record Set and ensures deleted keys behave
in the same way as non-existent keys. Future work will provide a way to
request entries by `entry-number`. This will allow tombstone entries to
be retrieved. We don't expect to support tombstones ever being visible
through a Record-style interface tho'. To see these keys, we would
expect the user to implement a Register Index in order to keep track of
which keys have been used in the past but are no longer in use. This has
the advantage that it would not only be possible to see this
information, but the user could write Index Functions that could use it
to enforce specific behaviour such as forbidding keys to be reused once
they had been tombstoned, forbidding keys to be changed once set,
ensuring that keys obey particular formatting or style guidelines or any
other domain specific policy that the user has a requirement for.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>