Rework register-root-digest
We're no longer backed by a Merkle Tree and therefore we don't have a
way to compute the root-digest anymore. For now we just need something
that changes as the Register changes so here we fake something up by
using the version number and some other data to make it convincingly
long.

We also mix in the backing-store-ref. This ensures that root-digests for
all versions of all Registers are always different. This seems safer
than always having identical root-digests for Registers that just happen
to have the same number of entrys but are otherwise entirely different.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>