SQLite Infrastructure
Here we introduce some infrastructure code for interacting with SQLite.
This will be used to persist Registers, Items and Entrys.

This patch supplies an initialisation routine `initialise-backing-store`
that allows the user to create the schema in the database before first
use. For now we hard-code the table names. This means that it is not
possible to have more than one Backing Store in a given SQLite database
and the user must ensure that those table names are not used by any
other users of the database.

We only support persistent tables: we don't yet support storing the
tables in SQLite's memory or temporary facilities.

We've taken a punt on some of the indexes that might be required. Of
particular interest is `entrys-region-key-entry-number-log-id`. One
might expect the log-id to appear earlier in the key hierarchy but it
should be useful as-is for record based queries (i.e. by region and key)
and it might also work if we wanted to find where common keys are
mentioned across a number of Registers. Who knows? Let's experiment!

The `registers` table has been structured so that Registers can be
designated as an `index-of` another Register. This will allow us to
implement Register Indexes as first class Register objects that can be
manipulated just like any other Register.

We also provide facilities for representing and running database
queries. This allows queries to be specified in terms of SQL query text,
serialisers for converting CHICKEN data structures to arguments suitable
for binding to the query and deserialiers for taking the columns
returned in the result set and converting them into data structures
suitable for CHICKEN.

We introduce a parameter db-ctx that holds the database handle. However,
because we only support a single handle, this is not yet threadsafe.
Further work will introduce a mechanism for the user to manage a pool of
database handles.

This patch hardcodes the name of the database file to
`orc.backing-store.sqlite`. This file ends up in the current directory.
Whilst we're a little rude, SQLite is pretty good at not stomping all
over things that aren't SQLite database files so the extent of the
badness is a little more than rudeness and a lot less than complete data
loss. (Things could get a little sticky if the file happens to contain
tables with the same name as tables that we use and they have a
compatible schemas)

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>