An SQLite backed Item store
This patch provides the item-store-ref, item-store-add! and
item-store-add-digest! internal procedures. These can be used by the
higher level Register code to store items and 'digest item-refs as well
as to retrieve items using a 'digest item-ref.

We don't currently use any transactions because we require SQLite's
SAVEPOINT transaction model and that is not currently available through
the sql-de-lite egg. This means that not only does it take a rather long
time to add a Register's worth of items, if something goes wrong in the
middle the database will be left with a load of dangling rows. This will
be fixed in future work.

When item-store-add! adds an item to the database it expects that item
to come with a 'digest item-ref. It will call onward to
item-store-add-digest! to have this item-ref added to the database. This
may not be what we want in the long (or even near)-term but it works for
now.

item-store-ref takes an item-ref and returns an item. For now we've only
implemented retrival of an item by 'digest item-refs as that's all we
need. However, it should be possible to use 'opaque item-refs as well.
item-store-ref should also be able to return *just* an item-ref rather
than a full item if it already knows about the item-ref but doesn't have
them item itself. This would allow us to support item declarations (such
as `declare-item` in RSF) that would enable redaction and "thin"
Registers.

If item-store-ref returns an item then it includes an 'opaque item-ref
as well.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>