Implement Item Scoping for read-rsf
Here we parameterize current-items for the duration of a read-rsf
operation.

This means that RSF files are processed in their own "scope" so that if
an `append-entry` statement refers to an item then the RSF file must
have previously supplied that item with an `add-item` statement. This
allow us to ensure that RSF files we process are self-consistent no
matter what we already have in the database or what calls to
`register-add-item` and `register-append-entry` the user has
independently performed.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>