Teach register-add-item about the SQLite Backing Store
Rework current-items so that it is an index that maps back and forth
between 'digest item-refs and 'opaque item-refs. This allows us to keep
track of which items we've seen no matter where the item-refs come from.

Rework register-add-item to tell the Backing Store about new items that
it sees and to use the current-items list to manage the items that have
been declared in the current "scope". If, within a given "scope", an
item is passed more than once to register-add-item, the current-items
code will emit a warning message on stderr.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>