Teach register-{records,record-ref} to use entry-store-{keys,key-ref}
Get Records from the Backing Store!

We also change the argument ordering of the
register-{records,record-ref} procedures to be more consistent with the
rest of the API by putting the register object first. This is less
consistent than some other Scheme APIs (such as alist-* and friends)
where the object to be operated on is of often the last non-optional
argument.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>