Expose key serialisation and deserialisation routines
string->key and key->string

These are the procedures from the Backing Store layer but it's useful for
the user to be able to serialise and deserialise keys as well.

The serialiser converts a key to a string with nice lexical semantics. The
lexical semantics are such that the returned string sorts in the order that
one would expect the original keys to sort in.

The user should treat the serialised string as completely opaque apart from
it's comparison properties.

The deserialiser converts the string with nice lexical semantics back to a
key.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>