Teach the Backing Store how to find Registers
This finds Registers or Indexes (which are also Registers) by name in the
Backing Store. To find Registers, the index-of argument should be 0. To
find named Indexes, the index-of argument should be the log-id of the
Register that it is an Index on. This means that Indexes can only be found
after their corresponding Registers have been found.

When we add support for using (and creating) Indexes we should probably
rework `register-store-ref` to work more like the `entry-store-key`*
procedures wherein the names of all the immediate child Indexes for the
given Register will also be returned.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>