lambda-in-savepoint: Wrap backing store calls in database transactions
Here we provide a macro, lambda-in-savepoint, that can be used in place
of (lambda ...) to ensure that the dynamic scope of the defined
procedure is run inside a database transaction. This macro is for the
sole use of the backing store procedures.

This requires our extensions to the sql-de-lite egg that exposes
SQLite's SAVEPOINT style transactions. SAVEPOINT transactions are
required here because they can nest and SQLite's regular transactions
cannot.

The use of transactions at the interface between the backing store and
the rest of the code ensures that individual backing store operations
are atomic and therefore correct.

Whilst backing store procedures do not call other backing store
procedures, we require that they use nestable transactions so that we
can later wrap a user's API call in its own transaction. This will
ensure that the API call as a whole is atomic and therefore correct even
if it uses multiple backing store operations.

A future patch will transactionalise the user facing APIs so, for now,
lambda-in-savepoint does not enforce the requirement that it be called
inside an existing transaction.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>