item-ref-opaque-evict: a quick and dirty way to allow item-refs to be used in applications
Evict an opaque item-ref so that users can put them in things like web
forms and then compare them to other item-refs in a POST handler. The
item-id is supposed to be entirely internal to the Backing Store and is
not guranteed to be stable. Therefore we deliberately introduce a bit of
instability so that users of the API don't come to rely on them.

Future work will more fully address the issues with item-refs. For now,
opaque item-refs allow us to keep the low level Register structure
independent of cryptographic concerns such as digest algorithms. In the
mean time, users need to be able to let users interact with items and
then work out what they did when they have finished.

There is no way to "unevict" an opaque item-ref. When you have an
evicted item-ref you can find it in an Entry's item list by evicting
those item-refs and then comparing them to what you have in hand. This
ensures that the item-ref data path is never user controlled and
therefore we don't have to be too defensive when handling them
internally. This allows us to provide item-ref-opqaue-evict quickly and
easily without departing too far from the original design goals and
constraints of item-refs.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>