define-in-transaction: Transactionalise API
Here we introduce the define-in-transaction macro that can can be used
in place of (define ...). This is used to wrap user facing API calls to
ensure that the whiole user operation is run inside a database
trasaction and is therefore atomic and correct.

Wrapping the entire operation in a transaction also means that, in most
cases, SQLite will not do expensive and time-consuming operations every
time transactions are used internally by orc to isolate different parts
of the work. This helps keep things fast as well as safe.

define-in-transaction is complicated by the fact that it needs to
distingush between user calls to APIs and internal calls to APIs.

When a user calls an API, define-in-transaction uses a standard,
non-nestable, SQLite transaction to ensure that the database is fully
committed at the point the API call returns. This ensures that the
user's operation is durable and will not be unexpectedly rolled back:
once we tell the user that it's complete, it really is comitted to the
backing store.

When orc makes an internal call to an API, define-in-transaction uses a
SAVEPOINT transaction. This means that the call is isolated from the
rest of the work but can (and must) also be nested inside the
transaction of the user's original call.

Finally, we install logic in lambda-in-transaction to ensure that its
documented sematics of requiring an enclosing transaction are
enforced.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>