Allow current-items-update! to accept an opaque item
In any given scope, items can be first seen during an API call by the
user or by reading them directly from the backing store. If we get them
from the user (via an add-item call) they end up with an item-ref of
'digest type. If we read them from the backing store they end up with an
item-ref of 'opaque type.

We need to be able to add the item-refs of 'opaque type to the current
scope so that if the user passes them back to us unchanged (for example,
as part of a multi-item entry or as an item for an entry with a new
key), we recognise them without the user having to do an explicit
add-item for something that is already in the backing store.

This patch changes current-items-update!: a future patch will ensure
that items are actually updated at all the right times.

current-items-update! should take an item-ref rather than an item so
that we can actually implement declare-item. For now we take an item
instead so that we can print nice error messages when we need to.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>