Ensure any items we return from the database are available in the current scope
...so that the user doesn't have to add-item when they want to read,
modify and write an entry with multiple items, some of which don't
change.

Besides, the user would find that add-item would not work because
add-item won't accept an item with an opaque ref!

We only call current-items-ref if the item's item-ref is not already in
the current scope. This is because we only have an item-ref of 'opaque
type and there might already be an item-ref of 'digest type associated
with it that we don't know about: having an item-ref of both types for
the same item in the scope is better than having an item-ref of 'opaque
type only.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>