Refactor backing store initialisation to be more amenable to user managed db handles
`initialise-backing-store` took a filename that it used to create a
database handle for itself. This database handle was thrown away at the end
of the procedure.

This meant that in memory or temporary databases could not be effectively
initialised because they are handle specific.

A side effect of this factorisation was that we never supplied
`open-backing-store`;  a procedure that would allow a user to
independently open their own database files and manage their own database
handles.

This patch cleans things up and paves the way for us to implement user
managed pools of database connections which will allow us to support multi-
threaded applications.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>