Supply `with-backing-store` to allow an application to specify the database handle to be used
`(db-ctx)` is now initialised to #f which lifts the restriction of the
backing store always being stored in ./orc.backing-store.sqlite.

However, we now require the user to
 + call `open-backing-store` themselves,
 + save the returned handle,
 + wrap their calls to the orc API in `with-backing-store`, remembering to
   supply the handle from `open-backing-store`.

This is slightly more complicated but allows the user to maintain a pool of
backing store handles so that their application can be multi-threaded.

The user must still ensure that `initialise-backing-store` is only ever
called once per file. Failure to do so will result in an exception being
thrown but will not damage or lose any state.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>