stream-query: Support streaming reads from the database
This patch provides `stream-query`: similar in functionality to
`run-query` but it allows the user to consume and process as much or as
little of the result set as and when they need it rather than doing it
all up front and returning everything to the caller.

i.e. It lets us have items one at a time when we want them rather than
processing the entire result set in one batch whilst still providing all
the bells and whistles of `run-query` such as the serialisers,
deserialisers and constructors.

It will allow us to efficiently use the merkle-tree egg as that expects
an interface that allows it to get the data entry-by-entry but its data
access is usually linear so we should often be able to reuse a running
statement when we want "the next" entry. This patch also constitutes
some of the groundwork that will allow us to implement more flexible
user facing APIs that, for example, return the data in chunks or allow
the user to sip on an iterator.

We achieve this by using call/cc to jump in and out of the control flow
of sql-de-lite's `query` interface. This could also be achieved without
call/cc by using sql-de-lite's lower level interface but I wanted to
continue to take advantage of sql-de-lite's statement management stuff.

`stream-query`'s interface is quite raw, providing a cleanup procedure
and an iterator that will return the next row. For the purposes of the
interface, we specify that callers must always call cleanup when they
are finished whether they exhaust the entire result set or not. It may
currently be permissable to avoid a call to cleanup in the event that
all the rows are exhausted, but we reserve the right to change the
implementation so that this is not the case in the future. Calling
cleanup once, however you finish, will always be safe and correct.

Finally, the cleanup and iterator procedures provided to the caller wrap
a running sql-de-lite statement and therefore cannot be passed out of
the scope of the current transaction. For the purpose of the interface
we specify that the procedures may not be passed across a transaction
boundary at all but the current implementation provided by this patch,
and the current state of sql-de-lite, mean that you can probably get
away with passing them down the tranasction stack but not up. However,
we reserve the right to change the implementation so that this is not
the case in the future.

We do support having multiple streaming queries in progress from the
same transaction. If this does not work then it is considered a bug,
however, I have not yet tested it thoroughly.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>