wide-entry-item-stream->entry-stream
...which is similar to `wide-entry-item-rows->entries` but supports
passing one entry at a time, along with its entry number, to the caller
over a (cleanup,iterator)-style interface.

This patch provides a high-level driver for use with `stream-query`,
similar to the existing `wide-entry-item-rows->entries` procedure that
is used with `run-query`. It executes the relevant query, keeps track of
the inter-row state, consumes database rows as necessary and passes one
entry at a time to the caller.

It is more general than `wide-entry-item-rows->entries` because it
allows the way that items are built up to be customised through its
`make-item` argument. Moreover, it is entry-number oriented rather than
key oriented so it can be used in situations that expect a stream of
entries rather than a stream of records.

When `wide-entry-item-stream->entry-stream` is used with `stream-query`
the same caveats regarding transaction boundaries and the necessity to
call `cleanup` apply.

Further work will eliminate `wide-entry-item-rows->entries` by
refactoring the places where it is used in terms of
`wide-entry-item-stream->entry-stream`.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>