entry-store-stream: Data access to get a block of entries out in order
Here we use `with-entry-item-stream->entry-stream` to get a block of
entries out in entry-number order (rather than the entry-key oriented
access we have used in the past).

The intention is to use this to integrate with the merkle-tree egg so
that we can calculate a root-digest for a Register. We will set up a
query to scan the entry space and then expect the root-digest
calculation process to access the leaves in order and therefore be able
to call the iterator to get the next entry from the running query rather
than have to start up a whole new query each time the merkle-tree egg
wants to do data access.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>