register-root-digest: A quick and dirty integration with the merkle-tree egg
This improves the status quo in so far as we do now get a root digest
that is a cryptographic function of the entry log rather than something
fake that just happens to be different for each register in a particular
backing store.

This will allow us to do the Open Permits Technology Demo.

However, it has a number of shortcomings:

 + We don't even take advantage of the streaming features of
   entry-store-stream; we just tear it down after each leaf.  It is,
   overall, more efficient to use a stream-query version rather than a
   run-query version tho'.

 + Quick and easy serialisation of entries. We just call `->string` on
   them without any attempt to canonicalise them,  or to ensure they are
   robust, secure or portable. In particular, the digest that is
   produced is not comparable with the GOV.UK Registers root digest.

Further work will address these issues and, until then,
`register-root-digest` is still not recommended for general use.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>