entry-audit-path to compute the proof for an entry
This is key based, rather than entry number based for now. Further work
will make it entry number based.

This also changes the entry-stream method to return
an opaque digest instead of a real one so that both
merkle trees are hashing the same content. Really we
don't want this and should revisit how we do refs
generally.

`entry-audit-path` signals an Error when the specified key does not
exist in the Register. This has been recorded in ticket #46.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>
Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>