Export register-records-range
We realised that we weren't sure what users expect from this API and, as
we weren't exporting it to them, we had no way to find out.

With the new streaming query infrastructure, we should be able to
implement this more efficiently because it still reads the entire result
set when the user calls it. Therefore, we expect this interface to
change substantially in the future, perhaps in the direction of
something more cursor or iterator based.

For now, it's available and we'll see how it gets used.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>