Add a simple command line interface
This commit adds a command-line executable (also called `orc`).
The `orc` tool will be used to provide command-line manipulation
of Registers using the library, and be a lower barrier to entry
for developers wanting to use the Registers ecosystem.

`orc` can be invoked with `-h` or `--help` to print usage.

== Installing
`orc` will be installed with the library when building the `orc`
egg using `chicken-install` from this directory,
`chicken-install orc` from elsewhere, or when installing `orc` as
an egg dependency of another application.

== Backing stores
Like the library, the CLI uses a backing store to store Registers.
Registers will first need to be read into the backing store before
they can be manipulated. Using a backing store allows `orc` to
provide its various features in a performant manner.

Backing stores should be invisible to most users. The key piece of
knowledge is that `orc` will use "./orc.register-store" as a
backing store by default, and will create this file if it doesn't
exist. This means that any Registers loaded and manipulated are
directory-local by default. Choosing a different backing store
can be done using the `-S` option to `orc`.

Signed-off-by: Simon Worthington <simon@register-dynamics.co.uk>