Fixup entry-audit-path to use entry numbers and add entry-audit-path* that is key-based.
`entry-audit-path` signals an Error when passed an entry number that is
greater than the version number of the current Register. This has been
recorded in ticket #47.

`entry-audit-path*` still has the problem documented in the previous
patch (and ticket #46) but it now manifests as an assertion failure when
the invalid entry number is passed to `entry-audit-path`.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>