skeleton-encode: serialisation routines for Register Key data
This is the first cut of the skeleton emitter.
It can emit all types, including user defined types but it's a bit lax
with the characters it allows in strings.

skeleton is a key encoding that can represent a range of primitive and
user defined types. The encoded data is always valid UTF-8 that sorts in
the same order as you would expect the unencoded data to sort in its
native type. i.e. numbers sort in numerical order rather than in string
prefix order. The numbers model allows for exact representation of
rational numbers of unlimited size and precision.

skeleton provides serialisers and deserialisers that can deal with the
low level encoding scheme that ensures Register Key data always sorts in
the correct order for its type across all Regiser implementations,
independently of platform, language or backing store. Moreover, the
encoded keys always consist only of printable characters which makes
them suitable for embedding in other formats and protocols.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>