entry-store-keys query improvement
`register-records` uses `entry-store-keys` to fetch the latest Entry for
each Key from the database.

For a Backing Store of ~49MiB that contained several Registers including
one with 42,737 Records in the user Region, it was taking ~150s for
`register-records` to return all those Records from the large Register:

-----
(time
  (length
    (with-backing-store (open-backing-store "qp.sqlite") (lambda ()
      (register-records r 'user)))))

->

159.196s CPU time, 0.18s GC time (major), 1218247/263115 mutations (total/tracked), 7/2165 GCs (major/minor), maximum live heap: 33.09 MiB
-----

With this patch, the time improves to less than 1s:

-----
0.704s CPU time, 0.148s GC time (major), 1218247/263088 mutations (total/tracked), 5/2167 GCs (major/minor), maximum live heap: 33.09 MiB
42737
-----

Most of the time was spent processing the query in SQLite. Therefore,
this patch:

  + Changes LEFT OUTER JOINs to INNER JOINs where possible.

  + Adds extra conditions to JOIN clauses so that the query planner can
    make better use of indexes.

  + Introduces missing ORDER BY clauses. The previous, seemingly
    correct, ordering was an artefact of the GROUP BY clause in the
    subselect.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>