Remove `wide-entry-item-rows->entries`
This code has been superceded by the more general and more powerful
`wide-entry-item-stream->entry-stream`.

As there are no longer any active uses for it and no reasons to use it
instead of its replacement, this patch removes it from the codebase.

It doesn't appear to have any dependencies that will be orphaned.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>