entry-store-key-ref query improvement
`entry-store-key-ref`, used by `records-range`, suffers from the same
problem as `entry-store-keys` so this patch introduces similar
changes.

`records-range` uses `entry-store-key-ref` to fetch the latest Entry for
each Key in the specified range from the database.

For a Backing Store of ~49MiB that contained several Registers including
one with 42,737 Records in the user Region, it was taking ~18s for
`register-records-range` to return slightly more than 10% of the Records
from the large Register:

-----
(time
  (length
    (with-backing-store (open-backing-store "qp.sqlite") (lambda ()
      (register-records-range r 'user (make-key "140001") (make-key "145943"))))))

->

18.168s CPU time, 146165/31548 mutations (total/tracked), 0/261 GCs (major/minor), maximum live heap: 8.3 MiB
5108
-----

This patch improves the performance to less than 0.1s thus:

-----
0.088s CPU time, 146153/31502 mutations (total/tracked), 0/261 GCs (major/minor), maximum live heap: 8.3 MiB
----

Most of the time was spent processing the query in SQLite. Therefore,
this patch:

  + Changes LEFT OUTER JOINs to INNER JOINs where possible.

  + Adds extra conditions to JOIN clauses so that the query planner can
    make better use of indexes.

  + Introduces missing ORDER BY clauses. The previous, seemingly
    correct, ordering was an artefact of the GROUP BY clause in the
    subselect.

Signed-off-by: Andy Bennett <andyjpb@register-dynamics.co.uk>